+++ src/fsd/1-pages/input-guild-roster-snapshots/roster-snapshots-tab.tsx
@@ -460,7 +460,13 @@ export const RosterSnapshotsTab = ({ members, memberStates, onLoadMembers }: Ros
             return;
         }
 
-        if (saveData?.status === 'TIMED_OUT') {
+        if (saveData === undefined) {
+            setSaveError('Request was canceled or returned no data.');
+            setSaveStage('error');
+            return;
+        }
+
+        if (saveData.status === 'TIMED_OUT') {
             setSaveError('The server timed out while saving the snapshot. Please try again.');
             setSaveStage('error');
             return;
